Extract rendering logic
